falta debug

+++ classes/Channel.cpp
@@ -7,7 +7,7 @@ Channel::Channel(std::string ChannelName)
 	_password = "";
 	_invite = false;
 	_clearanceTopic = false;
-	_clientSize = INT_MAX;
+	_clientSize = ULONG_MAX;
 }
 
 Channel::~Channel()
@@ -112,11 +112,11 @@ std::vector<Client>	Channel::getAdmins()
 	return (_admins);
 }
 
-int	Channel::getClientSize()
+unsigned long int	Channel::getClientSize()
 {
 	return (_clientSize);
 }
-void	Channel::setClientSize(int new_size)
+void	Channel::setClientSize(unsigned long int new_size)
 {
 	_clientSize = new_size;
 }
